+++ src/gui/pages/page0.py
@@ -1,0 +1,14 @@
+import flet as ft
+from .base import BasePage
+
+class Page0(BasePage):
+    def __init__(self):
+        super().__init__()
+        self._page = self.pinit()
+
+    def pinit(self):
+        return ft.Column(
+            controls=[ft.Text("Template"), self.bottom_bar],
+            alignment=ft.MainAxisAlignment.SPACE_BETWEEN,
+        )
+
