+++ src/gui/pages/page2.py
@@ -1,0 +1,63 @@
+import flet as ft
+from .base import TaskBasePage
+
+from tests import Task2
+
+
+class Page2(TaskBasePage):
+    def __init__(self):
+        super().__init__()
+        self.index = 2
+        self.data = {}
+        self.test = Task2()
+
+        self.num = ft.TextField(label="Number")
+        self.baseX = ft.TextField(label="Base x")
+        self.baseAns = ft.TextField(label="Base ans")
+        self.lgX = ft.TextField(label="Lg x")
+        self.lgAns = ft.TextField(label="Lg ans")
+
+        self.len_text = ft.Text("Fract len", weight=ft.FontWeight.BOLD)
+        self.res_text = ft.Text("Result", weight=ft.FontWeight.BOLD)
+        self.res_text.spans.append(ft.TextSpan(style=ft.TextStyle(size=9)))
+
+        self._page = self.pinit()
+
+    def pinit(self):
+        row = ft.Row
+        res_row = row(
+            controls=[self.len_text, self.evaluate_btn],
+            alignment=ft.MainAxisAlignment.SPACE_BETWEEN,
+        )
+
+        task_content = [
+            self.num,
+            self.baseX,
+            self.baseAns,
+            self.lgX,
+            self.lgAns,
+            ft.Divider(height=1),
+            res_row,
+            self.res_text
+        ]
+        top_part = self.join_top(task_content)
+        return self.join_page(top_part)
+
+    # ------------------------------------------------------------------------------------------------------
+
+    def process(self, e):
+        self.read({
+            "x": self.num, "base_x": self.baseX, "base_ans": self.baseAns,
+            "lg_x": self.lgX, "lg_ans": self.lgAns
+        })
+        assert self.check()
+        self.data["base_x"] = int(self.data["base_x"])
+        self.data["base_ans"] = int(self.data["base_ans"])
+        self.data["lg_x"] = float(self.data["lg_x"])
+        self.data["lg_ans"] = float(self.data["lg_ans"])
+        res = self.test.process(*self.data.values())
+        self.len_text.value = "Fract len {0} -> {1}".format(res[0], res[1])
+        self.res_text.value = "Result {0}".format(res[3][0])
+        self.res_text.spans[0].text = "{0}".format(res[2])
+        self._page.update()
+
